+++ src/DpressCliApp.php
@@ -8,6 +8,7 @@ use Dynart\Micro\CliOutputInterface;
 use Dynart\Micro\ConfigInterface;
 use Dynart\Micro\LoggerInterface;
 use Dynart\Micro\FormWidgets;
+use Dynart\Dpress\Content\Shortcodes;
 use Dynart\Micro\Micro;
 use Dynart\Micro\TranslationInterface;
 use Dynart\Micro\ViewInterface;
@@ -304,6 +305,7 @@ class DpressCliApp extends CliApp {
             Micro::get(TranslationInterface::class)
         );
         DpressServices::registerWidgets(Micro::get(FormWidgets::class));
+        DpressServices::registerShortcodes(Micro::get(Shortcodes::class));
         // there is no attribute processor in a CLI run, so the entities are registered by hand
         DpressServices::registerEntities(Micro::get(EntityManager::class));
         DpressServices::addMigrations(Micro::get(Migrations::class));